fix progress in merge callback

+++ base/src/main/java/com/github/kubatatami/judonetworking/callbacks/MergeCallback.java
@@ -63,7 +63,7 @@ public class MergeCallback<T> {
         for (int value : progressMap.values()) {
             progress += value;
         }
-        return (int) ((float) progress / (float) requests);
+        return progress / requests;
     }
 
     public final void addSuccess() {
